Move notes out of a card & just below nav
