+++ app/views/races/_attachments.slim
@@ -1,6 +1,5 @@
-ul.list-group.list-group-flush
-  - race.attachments.each do |file|
-    li.list-group-item
-      = file.filename
-      a.ml-2 href=rails_blob_path(file, disposition: 'attachment')
-        = icon('fas', 'file-download')
+- race.attachments.each do |file|
+  li.list-group-item
+    = file.filename
+    a.ml-2 href=rails_blob_path(file, disposition: 'attachment')
+      = icon('fas', 'file-download')
